+++ libtelemetry/src/test/java/com/mapbox/android/telemetry/LocationReceiverTest.java
@@ -5,9 +5,12 @@ import android.content.Context;
 import android.content.Intent;
 import android.location.Location;
 import android.location.LocationManager;
-import android.os.Bundle;
 
+import android.os.Parcelable;
 import org.junit.Test;
+
+import java.util.ArrayList;
+import java.util.Collections;
 
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.eq;
@@ -19,240 +22,140 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 public class LocationReceiverTest {
+  private enum Field { Lat, Lon, Alt; }
+
+  private static final int NUMBER_OF_LOCATIONS = 5;
+
+  private static Location[] invalidLocations = {getInvalidLocation(Field.Lat, Double.NaN),
+          getInvalidLocation(Field.Lon, Double.NaN),
+          getInvalidLocation(Field.Alt, Double.NaN),
+          getInvalidLocation(Float.NaN)};
+  private static Location[] positiveInfLocations = {getInvalidLocation(Field.Lat, Double.POSITIVE_INFINITY),
+          getInvalidLocation(Field.Lon, Double.POSITIVE_INFINITY),
+          getInvalidLocation(Field.Alt, Double.POSITIVE_INFINITY),
+          getInvalidLocation(Float.POSITIVE_INFINITY)};
+  private static Location[] negativeInfLocations = {getInvalidLocation(Field.Lat, Double.NEGATIVE_INFINITY),
+          getInvalidLocation(Field.Lon, Double.NEGATIVE_INFINITY),
+          getInvalidLocation(Field.Alt, Double.NEGATIVE_INFINITY),
+          getInvalidLocation(Float.NEGATIVE_INFINITY)};
 
   @Test
-  public void checksSendEventCalled() throws Exception {
+  public void checksSendEventCalled() {
     Context mockedContext = mock(Context.class);
     MapboxTelemetry.applicationContext = mockedContext;
     Intent mockedIntent = mock(Intent.class);
     when(mockedIntent.getStringExtra(eq("location_received"))).thenReturn("onLocation");
     ActivityManager mockedActivityManager = mock(ActivityManager.class, RETURNS_DEEP_STUBS);
     when(mockedContext.getSystemService(Context.ACTIVITY_SERVICE)).thenReturn(mockedActivityManager);
-    Bundle mockedBundle = mock(Bundle.class);
-    when(mockedIntent.getExtras()).thenReturn(mockedBundle);
-    Location mockedLocation = mock(Location.class);
-    when(mockedBundle.get(eq(LocationManager.KEY_LOCATION_CHANGED))).thenReturn(mockedLocation);
+
+    ArrayList<Parcelable> mockedLocations = getLocations(NUMBER_OF_LOCATIONS);
+    when(mockedIntent.getParcelableArrayListExtra(eq(LocationManager.KEY_LOCATION_CHANGED)))
+            .thenReturn(mockedLocations);
     EventCallback mockedEventCallback = mock(EventCallback.class);
     LocationReceiver theLocationReceiver = new LocationReceiver(mockedEventCallback);
 
     theLocationReceiver.onReceive(mockedContext, mockedIntent);
-
-    verify(mockedEventCallback, times(1)).onEventReceived(any(LocationEvent.class));
+    verify(mockedEventCallback, times(NUMBER_OF_LOCATIONS)).onEventReceived(any(LocationEvent.class));
   }
 
   @Test
-  public void checksSendEventNotCalledWhenLatitudeNaN() throws Exception {
+  public void checksSendEventNotCalledWhenLocationNaN() {
     Context mockedContext = mock(Context.class);
     Intent mockedIntent = mock(Intent.class);
     when(mockedIntent.getStringExtra(eq("location_received"))).thenReturn("onLocation");
-    Bundle mockedBundle = mock(Bundle.class);
-    when(mockedIntent.getExtras()).thenReturn(mockedBundle);
-    Location mockedLocation = mock(Location.class);
-    when(mockedBundle.get(eq(LocationManager.KEY_LOCATION_CHANGED))).thenReturn(mockedLocation);
-    when(mockedLocation.getLatitude()).thenReturn(Double.NaN);
+
+    MapboxTelemetry.applicationContext = mockedContext;
+    ActivityManager mockedActivityManager = mock(ActivityManager.class, RETURNS_DEEP_STUBS);
+    when(mockedContext.getSystemService(Context.ACTIVITY_SERVICE)).thenReturn(mockedActivityManager);
+
+    ArrayList<Parcelable> mockedLocations = getInvalidLocations(invalidLocations);
+    when(mockedIntent.getParcelableArrayListExtra(eq(LocationManager.KEY_LOCATION_CHANGED)))
+            .thenReturn(mockedLocations);
+
     EventCallback mockedEventCallback = mock(EventCallback.class);
     LocationReceiver theLocationReceiver = new LocationReceiver(mockedEventCallback);
 
     theLocationReceiver.onReceive(mockedContext, mockedIntent);
-
     verify(mockedEventCallback, never()).onEventReceived(any(LocationEvent.class));
   }
 
   @Test
-  public void checksSendEventNotCalledWhenLongitudeNaN() throws Exception {
+  public void checksSendEventNotCalledWhenLocationPositiveInfinite() {
     Context mockedContext = mock(Context.class);
     Intent mockedIntent = mock(Intent.class);
     when(mockedIntent.getStringExtra(eq("location_received"))).thenReturn("onLocation");
-    Bundle mockedBundle = mock(Bundle.class);
-    when(mockedIntent.getExtras()).thenReturn(mockedBundle);
-    Location mockedLocation = mock(Location.class);
-    when(mockedBundle.get(eq(LocationManager.KEY_LOCATION_CHANGED))).thenReturn(mockedLocation);
-    when(mockedLocation.getLongitude()).thenReturn(Double.NaN);
+
+    MapboxTelemetry.applicationContext = mockedContext;
+    ActivityManager mockedActivityManager = mock(ActivityManager.class, RETURNS_DEEP_STUBS);
+    when(mockedContext.getSystemService(Context.ACTIVITY_SERVICE)).thenReturn(mockedActivityManager);
+
+    ArrayList<Parcelable> mockedLocations = getInvalidLocations(positiveInfLocations);
+    when(mockedIntent.getParcelableArrayListExtra(eq(LocationManager.KEY_LOCATION_CHANGED)))
+            .thenReturn(mockedLocations);
+
     EventCallback mockedEventCallback = mock(EventCallback.class);
     LocationReceiver theLocationReceiver = new LocationReceiver(mockedEventCallback);
 
     theLocationReceiver.onReceive(mockedContext, mockedIntent);
-
     verify(mockedEventCallback, never()).onEventReceived(any(LocationEvent.class));
   }
 
   @Test
-  public void checksSendEventNotCalledWhenAltitudeNaN() throws Exception {
+  public void checksSendEventNotCalledWhenLocationNegativeInfinite() {
     Context mockedContext = mock(Context.class);
     Intent mockedIntent = mock(Intent.class);
     when(mockedIntent.getStringExtra(eq("location_received"))).thenReturn("onLocation");
-    Bundle mockedBundle = mock(Bundle.class);
-    when(mockedIntent.getExtras()).thenReturn(mockedBundle);
-    Location mockedLocation = mock(Location.class);
-    when(mockedBundle.get(eq(LocationManager.KEY_LOCATION_CHANGED))).thenReturn(mockedLocation);
-    when(mockedLocation.getAltitude()).thenReturn(Double.NaN);
+
+    MapboxTelemetry.applicationContext = mockedContext;
+    ActivityManager mockedActivityManager = mock(ActivityManager.class, RETURNS_DEEP_STUBS);
+    when(mockedContext.getSystemService(Context.ACTIVITY_SERVICE)).thenReturn(mockedActivityManager);
+
+    ArrayList<Parcelable> mockedLocations = getInvalidLocations(negativeInfLocations);
+    when(mockedIntent.getParcelableArrayListExtra(eq(LocationManager.KEY_LOCATION_CHANGED)))
+            .thenReturn(mockedLocations);
+
     EventCallback mockedEventCallback = mock(EventCallback.class);
     LocationReceiver theLocationReceiver = new LocationReceiver(mockedEventCallback);
 
     theLocationReceiver.onReceive(mockedContext, mockedIntent);
-
     verify(mockedEventCallback, never()).onEventReceived(any(LocationEvent.class));
   }
 
-  @Test
-  public void checksSendEventNotCalledWhenAccuracyNaN() throws Exception {
-    Context mockedContext = mock(Context.class);
-    Intent mockedIntent = mock(Intent.class);
-    when(mockedIntent.getStringExtra(eq("location_received"))).thenReturn("onLocation");
-    Bundle mockedBundle = mock(Bundle.class);
-    when(mockedIntent.getExtras()).thenReturn(mockedBundle);
-    Location mockedLocation = mock(Location.class);
-    when(mockedBundle.get(eq(LocationManager.KEY_LOCATION_CHANGED))).thenReturn(mockedLocation);
-    when(mockedLocation.getAccuracy()).thenReturn(Float.NaN);
-    EventCallback mockedEventCallback = mock(EventCallback.class);
-    LocationReceiver theLocationReceiver = new LocationReceiver(mockedEventCallback);
-
-    theLocationReceiver.onReceive(mockedContext, mockedIntent);
-
-    verify(mockedEventCallback, never()).onEventReceived(any(LocationEvent.class));
+  private static Location getInvalidLocation(Field index, double value) {
+    Location location = mock(Location.class);
+    switch (index) {
+      case Lat:
+        when(location.getLatitude()).thenReturn(value);
+        break;
+      case Lon:
+        when(location.getLongitude()).thenReturn(value);
+        break;
+      case Alt:
+        when(location.getAltitude()).thenReturn(value);
+        break;
+      default:
+        break;
+    }
+    return location;
   }
 
-  @Test
-  public void checksSendEventNotCalledWhenPositiveInfiniteLatitude() throws Exception {
-    Context mockedContext = mock(Context.class);
-    Intent mockedIntent = mock(Intent.class);
-    when(mockedIntent.getStringExtra(eq("location_received"))).thenReturn("onLocation");
-    Bundle mockedBundle = mock(Bundle.class);
-    when(mockedIntent.getExtras()).thenReturn(mockedBundle);
-    Location mockedLocation = mock(Location.class);
-    when(mockedBundle.get(eq(LocationManager.KEY_LOCATION_CHANGED))).thenReturn(mockedLocation);
-    when(mockedLocation.getLatitude()).thenReturn(Double.POSITIVE_INFINITY);
-    EventCallback mockedEventCallback = mock(EventCallback.class);
-    LocationReceiver theLocationReceiver = new LocationReceiver(mockedEventCallback);
-
-    theLocationReceiver.onReceive(mockedContext, mockedIntent);
-
-    verify(mockedEventCallback, never()).onEventReceived(any(LocationEvent.class));
+  private static Location getInvalidLocation(float value) {
+    Location location = mock(Location.class);
+    when(location.getAccuracy()).thenReturn(value);
+    return location;
   }
 
-  @Test
-  public void checksSendEventNotCalledWhenPositiveInfiniteLongitude() throws Exception {
-    Context mockedContext = mock(Context.class);
-    Intent mockedIntent = mock(Intent.class);
-    when(mockedIntent.getStringExtra(eq("location_received"))).thenReturn("onLocation");
-    Bundle mockedBundle = mock(Bundle.class);
-    when(mockedIntent.getExtras()).thenReturn(mockedBundle);
-    Location mockedLocation = mock(Location.class);
-    when(mockedBundle.get(eq(LocationManager.KEY_LOCATION_CHANGED))).thenReturn(mockedLocation);
-    when(mockedLocation.getLongitude()).thenReturn(Double.POSITIVE_INFINITY);
-    EventCallback mockedEventCallback = mock(EventCallback.class);
-    LocationReceiver theLocationReceiver = new LocationReceiver(mockedEventCallback);
-
-    theLocationReceiver.onReceive(mockedContext, mockedIntent);
-
-    verify(mockedEventCallback, never()).onEventReceived(any(LocationEvent.class));
+  private static ArrayList<Parcelable> getInvalidLocations(Location[] locations) {
+    ArrayList<Parcelable> locationList = new ArrayList<>();
+    Collections.addAll(locationList, locations);
+    return locationList;
   }
 
-  @Test
-  public void checksSendEventNotCalledWhenPositiveInfiniteAltitude() throws Exception {
-    Context mockedContext = mock(Context.class);
-    Intent mockedIntent = mock(Intent.class);
-    when(mockedIntent.getStringExtra(eq("location_received"))).thenReturn("onLocation");
-    Bundle mockedBundle = mock(Bundle.class);
-    when(mockedIntent.getExtras()).thenReturn(mockedBundle);
-    Location mockedLocation = mock(Location.class);
-    when(mockedBundle.get(eq(LocationManager.KEY_LOCATION_CHANGED))).thenReturn(mockedLocation);
-    when(mockedLocation.getAltitude()).thenReturn(Double.POSITIVE_INFINITY);
-    EventCallback mockedEventCallback = mock(EventCallback.class);
-    LocationReceiver theLocationReceiver = new LocationReceiver(mockedEventCallback);
-
-    theLocationReceiver.onReceive(mockedContext, mockedIntent);
-
-    verify(mockedEventCallback, never()).onEventReceived(any(LocationEvent.class));
-  }
-
-  @Test
-  public void checksSendEventNotCalledWhenPositiveInfiniteAccuracy() throws Exception {
-    Context mockedContext = mock(Context.class);
-    Intent mockedIntent = mock(Intent.class);
-    when(mockedIntent.getStringExtra(eq("location_received"))).thenReturn("onLocation");
-    Bundle mockedBundle = mock(Bundle.class);
-    when(mockedIntent.getExtras()).thenReturn(mockedBundle);
-    Location mockedLocation = mock(Location.class);
-    when(mockedBundle.get(eq(LocationManager.KEY_LOCATION_CHANGED))).thenReturn(mockedLocation);
-    when(mockedLocation.getAccuracy()).thenReturn(Float.POSITIVE_INFINITY);
-    EventCallback mockedEventCallback = mock(EventCallback.class);
-    LocationReceiver theLocationReceiver = new LocationReceiver(mockedEventCallback);
-
-    theLocationReceiver.onReceive(mockedContext, mockedIntent);
-
-    verify(mockedEventCallback, never()).onEventReceived(any(LocationEvent.class));
-  }
-
-  @Test
-  public void checksSendEventNotCalledWhenNegativeInfiniteLatitude() throws Exception {
-    Context mockedContext = mock(Context.class);
-    Intent mockedIntent = mock(Intent.class);
-    when(mockedIntent.getStringExtra(eq("location_received"))).thenReturn("onLocation");
-    Bundle mockedBundle = mock(Bundle.class);
-    when(mockedIntent.getExtras()).thenReturn(mockedBundle);
-    Location mockedLocation = mock(Location.class);
-    when(mockedBundle.get(eq(LocationManager.KEY_LOCATION_CHANGED))).thenReturn(mockedLocation);
-    when(mockedLocation.getLatitude()).thenReturn(Double.NEGATIVE_INFINITY);
-    EventCallback mockedEventCallback = mock(EventCallback.class);
-    LocationReceiver theLocationReceiver = new LocationReceiver(mockedEventCallback);
-
-    theLocationReceiver.onReceive(mockedContext, mockedIntent);
-
-    verify(mockedEventCallback, never()).onEventReceived(any(LocationEvent.class));
-  }
-
-  @Test
-  public void checksSendEventNotCalledWhenNegativeInfiniteLongitude() throws Exception {
-    Context mockedContext = mock(Context.class);
-    Intent mockedIntent = mock(Intent.class);
-    when(mockedIntent.getStringExtra(eq("location_received"))).thenReturn("onLocation");
-    Bundle mockedBundle = mock(Bundle.class);
-    when(mockedIntent.getExtras()).thenReturn(mockedBundle);
-    Location mockedLocation = mock(Location.class);
-    when(mockedBundle.get(eq(LocationManager.KEY_LOCATION_CHANGED))).thenReturn(mockedLocation);
-    when(mockedLocation.getLongitude()).thenReturn(Double.NEGATIVE_INFINITY);
-    EventCallback mockedEventCallback = mock(EventCallback.class);
-    LocationReceiver theLocationReceiver = new LocationReceiver(mockedEventCallback);
-
-    theLocationReceiver.onReceive(mockedContext, mockedIntent);
-
-    verify(mockedEventCallback, never()).onEventReceived(any(LocationEvent.class));
-  }
-
-  @Test
-  public void checksSendEventNotCalledWhenNegativeInfiniteAltitude() throws Exception {
-    Context mockedContext = mock(Context.class);
-    Intent mockedIntent = mock(Intent.class);
-    when(mockedIntent.getStringExtra(eq("location_received"))).thenReturn("onLocation");
-    Bundle mockedBundle = mock(Bundle.class);
-    when(mockedIntent.getExtras()).thenReturn(mockedBundle);
-    Location mockedLocation = mock(Location.class);
-    when(mockedBundle.get(eq(LocationManager.KEY_LOCATION_CHANGED))).thenReturn(mockedLocation);
-    when(mockedLocation.getAltitude()).thenReturn(Double.NEGATIVE_INFINITY);
-    EventCallback mockedEventCallback = mock(EventCallback.class);
-    LocationReceiver theLocationReceiver = new LocationReceiver(mockedEventCallback);
-
-    theLocationReceiver.onReceive(mockedContext, mockedIntent);
-
-    verify(mockedEventCallback, never()).onEventReceived(any(LocationEvent.class));
-  }
-
-  @Test
-  public void checksSendEventNotCalledWhenNegativeInfiniteAccuracy() throws Exception {
-    Context mockedContext = mock(Context.class);
-    Intent mockedIntent = mock(Intent.class);
-    when(mockedIntent.getStringExtra(eq("location_received"))).thenReturn("onLocation");
-    Bundle mockedBundle = mock(Bundle.class);
-    when(mockedIntent.getExtras()).thenReturn(mockedBundle);
-    Location mockedLocation = mock(Location.class);
-    when(mockedBundle.get(eq(LocationManager.KEY_LOCATION_CHANGED))).thenReturn(mockedLocation);
-    when(mockedLocation.getAccuracy()).thenReturn(Float.NEGATIVE_INFINITY);
-    EventCallback mockedEventCallback = mock(EventCallback.class);
-    LocationReceiver theLocationReceiver = new LocationReceiver(mockedEventCallback);
-
-    theLocationReceiver.onReceive(mockedContext, mockedIntent);
-
-    verify(mockedEventCallback, never()).onEventReceived(any(LocationEvent.class));
+  private static ArrayList<Parcelable> getLocations(int mocksCount) {
+    ArrayList<Parcelable> locations = new ArrayList<>();
+    for (int i = 0; i < mocksCount; i++) {
+      locations.add(mock(Location.class));
+    }
+    return locations;
   }
 }